web/backend/requirements.txt
-r ../../scripts/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
PyYAML>=6.0
python-dotenv>=1.0.0
